Update bash command to remove a potential pid

+++ entrypoint.sh
@@ -2,10 +2,9 @@
 set -e
 
 # Remove a potentially pre-existing server.pid for Rails.
+DIR=/flix/tmp/pids
 FILE=/flix/tmp/pids/server.pid
-if [[ -f "$FILE" ]]; then
-    rm -f $FILE
-fi
+[ -d "$DIR" ] && [ -f "$FILE" ] && rm -f $FILE
 
 # Then exec the container's main process (what's set as CMD in the Dockerfile).
 exec "$@"
